complete Azure.Template to match DPG guideline
