+++ sdk/template/Azure.Template/.content/samples/README.md
@@ -11,4 +11,4 @@ description: Samples for the Azure.ServiceTemplate.Template client library.
 
 # Azure.ServiceTemplate.Template Samples
 
-- [Creating Resource](https://github.com/Azure/azure-sdk-for-net/blob/3ac301ac6435c818ad7a9946ab1c4023cee236ff/eng/templates/Azure.ServiceTemplate.Template/samples/Sample1_CreateResource.md)
+/* please refer to <https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/template/Azure.Template/samples/README.md> to write sample readme. */
