+++ sdk/template/Azure.Template/.content/samples/Sample1_HelloWorld.md
@@ -44,3 +44,5 @@ To see the full example source files, see:
 
 [README]: https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/anomalydetector/Azure.AI.AnomalyDetector/README.md
 [SampleData]: https://github.com/Azure/azure-sdk-for-net/tree/main/sdk/anomalydetector/Azure.AI.AnomalyDetector/tests/samples/data/request-data.csv
+
+/* please refer to <https://raw.githubusercontent.com/Azure/azure-sdk-for-net/main/sdk/template/Azure.Template/samples/Sample1_HelloWorld.md> to write sample readme file. */
