+++ sdk/template/Azure.Template/CHANGELOG.md
@@ -1,23 +1,12 @@
 # Release History
 
-## 1.0.0-beta.1 (Unreleased)
-
-### Breaking Changes
-- Breaking change 1
-- Breaking change 2
-- ...
+## 1.0.0-beta.1 (2022-04-13)
 
 ### Features Added
-- Feature 1
-- Feature 2
-- ...
+- Get secret
+
+### Breaking Changes
 
 ### Bugs Fixed
-- Bug fix 1
-- Bug fix 2
-- ...
 
 ### Other Changes
-- Change 1
-- Change 2
-- ...
