+++ sdk/template/Azure.Template/samples/Sample1.HelloWorldAsync.md
@@ -8,18 +8,18 @@ To use these samples, you'll first need to set up resources. See [getting starte
 using Azure.Template.Models;
 ```
 
-## Get secrets synchronously
+## Get secrets asynchronously
 
-You can create a client and get secrets synchronously:
+You can also get secrets asynchronously:
 
-```C# Snippet:Azure_Template_GetSecret
+```C# Snippet:Azure_Template_GetSecretAsync
 string endpoint = "https://myvault.vault.azure.net";
 var client = new MiniSecretClient(new Uri(endpoint), new DefaultAzureCredential());
 
-SecretBundle secret = client.GetSecret("TestSecret");
+SecretBundle secret = await client.GetSecretAsync("TestSecret");
 
 Console.WriteLine(secret.Value);
 ```
 
 To see the full example source files, see:
-* [HelloWorld](https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/template/Azure.Template/tests/Samples/Sample1_HelloWorld.cs)
+* [HelloWorldAsync](https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/template/Azure.Template/tests/Samples/Sample1_HelloWorldAsync.cs)
