+++ sdk/template/Azure.Template/tests/Samples/Sample1_HelloWorldAsync.cs
@@ -1,0 +1,37 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Threading.Tasks;
+using Azure.Core.TestFramework;
+using Azure.Identity;
+#region Snippet:Azure_Template_Namespaces
+using Azure.Template.Models;
+#endregion
+using NUnit.Framework;
+
+namespace Azure.Template.Tests.Samples
+{
+    public partial class Sample1_HelloWorld: SamplesBase<TemplateClientTestEnvironment>
+    {
+        [Test]
+        [AsyncOnly]
+        public async Task GettingASecretAsync()
+        {
+            var endpoint = TestEnvironment.KeyVaultUri;
+
+            #region Snippet:Azure_Template_GetSecretAsync
+#if SNIPPET
+            string endpoint = "https://myvault.vault.azure.net";
+#endif
+            var client = new TemplateClient(endpoint, new DefaultAzureCredential());
+
+            SecretBundle secret = await client.GetSecretValueAsync("TestSecret");
+
+            Console.WriteLine(secret.Value);
+            #endregion
+
+            Assert.NotNull(secret.Value);
+        }
+    }
+}
