+++ sdk/template/Azure.Template/tests/SessionRecords/TemplateClientLiveTests/CanGetSecret.json
