+++ sdk/template/Azure.Template/tests/SessionRecords/TemplateClientLiveTests/CanGetSecretAsync.json
